made player movable (no collision)
player represented by '@' , should change color for easier locating and adding collision to walls to prevent players from going oob.
